+++ Saver/Controller/CategoryDetail/DetailCategoryTransactionAmoutViewController.swift
@@ -57,18 +57,6 @@ class DetailCategoryTransactionAmoutViewController: UIViewController{
             categoryListTableView.bottomAnchor.constraint(equalTo: view.bottomAnchor)
         ])
     }
-    
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
 }
 
 extension DetailCategoryTransactionAmoutViewController: UITableViewDelegate, UITableViewDataSource{
