+++ Saver/UIFontExtension.swift
@@ -1,0 +1,29 @@
+//
+//  UIFontExtension.swift
+//  Saver
+//
+//  Created by 이상민 on 6/26/24.
+//
+
+import UIKit
+
+//MARK: - UIFont 커스텀
+extension UIFont{
+    //Title
+    static let saverTitleBold = UIFont.systemFont(ofSize: 25, weight: .bold)
+    static let saverTitleRegular = UIFont.systemFont(ofSize: 25, weight: .regular)
+    
+    //SubTitle
+    static let saverSubTitleSemibold = UIFont.systemFont(ofSize: 20, weight: .semibold)
+    static let saverSubTitleRegular = UIFont.systemFont(ofSize: 20, weight: .regular)
+    
+    //Body
+    static let saverBody1Semibold = UIFont.systemFont(ofSize: 16, weight: .semibold)
+    static let saverBody1Regurlar = UIFont.systemFont(ofSize: 16, weight: .regular)
+    static let saverBody2Semibold = UIFont.systemFont(ofSize: 14, weight: .semibold)
+    static let saverBody2Regurlar = UIFont.systemFont(ofSize: 14, weight: .regular)
+    
+    //Caption
+    static let saverCaption1Regular = UIFont.systemFont(ofSize: 12, weight: .regular)
+    static let saverCaption2Regular = UIFont.systemFont(ofSize: 10, weight: .regular)
+}
